Fix Werror in clang

+++ benchmark/peer_b.c
@@ -129,11 +129,11 @@ static void _recv_comm_packet(CommPacket *packet) {
     printf("[%d:ER%d] total=%ldKB, cost=%ld-%02ld:%02ld:%02ld, speed=%.2fKB/s, BW RATIO=%.2f%%\n",
            BAUD_RATE,
            err_cnt,
-           total_len >> 10,
-           cost / (3600 * 24),
-           cost % (3600 * 24) / 3600,
-           cost % (3600 * 24) % 3600 / 60,
-           cost % (3600 * 24) % 3600 % 60,
+           (long)total_len >> 10,
+           (long)cost / (3600 * 24),
+           (long)cost % (3600 * 24) / 3600,
+           (long)cost % (3600 * 24) % 3600 / 60,
+           (long)cost % (3600 * 24) % 3600 % 60,
            avg_speed, avg_speed / (BAUD_RATE >> 13) * 100);
     start = now;
   }
